--- packages/frontend/src/components/features/agents/agent-chat/agent-chat-session-runtime.ts
@@ -1,11 +1,0 @@
-import {
-  resolveAttachedSessionRuntimeQueryState,
-  type SessionRuntimeQueryInput,
-  type SessionRuntimeQueryState,
-} from "@/state/operations/agent-orchestrator/support/session-runtime-query-state";
-
-export { resolveAttachedSessionRuntimeQueryState };
-
-export type AgentChatSessionRuntimeQueryInput = SessionRuntimeQueryInput;
-
-export type AgentChatSessionRuntimeQueryState = SessionRuntimeQueryState;
