+++ packages/frontend/src/components/features/agents/agent-chat/use-agent-chat-session-runtime-data.ts
@@ -6,6 +6,7 @@ import {
   deriveAgentSessionViewLifecycle,
   type SessionRepoReadinessState,
 } from "@/state/operations/agent-orchestrator/lifecycle/session-view-lifecycle";
+import { resolveAttachedSessionRuntimeQueryState } from "@/state/operations/agent-orchestrator/support/session-runtime-query-state";
 import {
   SESSION_MODEL_CATALOG_STALE_TIME_MS,
   SESSION_TODOS_STALE_TIME_MS,
@@ -13,7 +14,6 @@ import {
   sessionTodosQueryOptions,
 } from "@/state/queries/agent-session-runtime";
 import type { AgentSessionState } from "@/types/agent-orchestrator";
-import { resolveAttachedSessionRuntimeQueryState } from "./agent-chat-session-runtime";
 
 type UseAgentChatSessionRuntimeDataArgs = {
   session: AgentSessionState | null;
