--- packages/frontend/src/pages/agents/agent-studio-session-runtime.ts
@@ -1,11 +1,0 @@
-import {
-  resolveAttachedSessionRuntimeQueryState,
-  type SessionRuntimeQueryInput,
-  type SessionRuntimeQueryState,
-} from "@/state/operations/agent-orchestrator/support/session-runtime-query-state";
-
-export { resolveAttachedSessionRuntimeQueryState };
-
-export type AgentStudioSessionRuntimeQueryInput = SessionRuntimeQueryInput;
-
-export type AgentStudioSessionRuntimeQueryState = SessionRuntimeQueryState;
